+++ src/components/controls/controls.vue
@@ -118,5 +118,4 @@ export default {
 };
 </script>
 
-<style lang="scss" src="./controls.scss" scoped>
-</style>
+<style lang="scss" src="./controls.scss" scoped />
