+++ src/components/inspectors/inspector.scss
