+++ src/components/miniPaper/miniPaper.scss
