+++ src/components/modeler/modeler.scss
@@ -1,3 +1,5 @@
+@import '../../../node_modules/jointjs/dist/joint.min.css';
+
 .modeler {
   .main-paper {
     position: absolute;
