+++ src/components/nodes/index.js
@@ -23,9 +23,9 @@ export { default as serviceTask } from './serviceTask';
 export { default as textAnnotation } from './textAnnotation';
 export { default as pool } from './pool';
 export { default as poolLane } from './poolLane';
-export { default as Statusbar } from '../Statusbar';
-export { default as ValidationStatus } from '../ValidationStatus';
-export { default as Modeler } from '../Modeler';
+export { default as Statusbar } from '../statusBar/Statusbar';
+export { default as ValidationStatus } from '../validationStatus/ValidationStatus';
+export { default as Modeler } from '../modeler/Modeler';
 export { default as manualTask } from './manualTask';
 export { default as boundaryTimerEvent } from './boundaryTimerEvent';
 export { default as boundaryMessageEvent } from './boundaryMessageEvent';
