+++ src/components/statusBar/statusBar.scss
