+++ src/components/toolbar/ToolBar.vue
@@ -166,3 +166,4 @@ export default {
   },
 };
 </script>
+<style lang="scss" src="./toolbar.scss" scoped />
