+++ src/components/validationStatus/validationStatus.scss
